requirements.txt
africastalking==1.2.9
amqp==5.3.1
asgiref==3.8.1
better-profanity==0.7.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
django-graphql-jwt==0.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
kombu==5.5.1
pillow==11.2.1
promise==2.3
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
schema==0.7.7
sightengine==1.4.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
